Revert "Add new folder structure for the new CAN commands"
