+++ .gitignore
@@ -1,6 +1,3 @@
 
 ~$GopherCAN.xlsx
 base_types.h
-gophercannon/reports/*
-gophercannon/outputs/*
-gophercannon/venv/*
