+++ GopherCAN_example.c
@@ -28,7 +28,7 @@ CAN_HandleTypeDef* example_hcan;
 
 // use this section to choose what module this should be (for testing 2 dev boards)
 // and what functionality should be enabled
-//#define F7XX_ENABLED                                                          // (F0xx otherwise)
+//#define F7XX                                                          // (F0xx otherwise)
 #define THIS_ACM
 //#define THIS_PDM
 //#define ENABLE_LOGIC_REQ
@@ -102,7 +102,7 @@ void init(CAN_HandleTypeDef* hcan_ptr)
 	float_tester.update_enabled = TRUE;
 
 	// adding can_callback_function
-#ifndef F7XX_ENABLED
+#ifndef F7XX
 	led_to_change = GPIO_PIN_5; // this is LD2 on the dev board
 #else
 	led_to_change = GPIO_PIN_0; // this is LD1 on the dev board
@@ -207,7 +207,7 @@ void main_loop()
 	button_state = !HAL_GPIO_ReadPin(GPIOC, GPIO_PIN_13);
 
 	// for some reason the button is the opposite on the F7
-#ifdef F7XX_ENABLED
+#ifdef F7XX
 	button_state = !button_state;
 #endif
 
@@ -238,7 +238,7 @@ void change_led_state(void* parameter, U8 remote_param)
 	// the LED to change is dependent on the parameter stored on this module (*((U16*)parameter))
 	U16 gpio_pin = *((U16*)parameter);
 
-#ifndef F7XX_ENABLED
+#ifndef F7XX
 	HAL_GPIO_WritePin(GPIOA, gpio_pin, !!remote_param);
 #else
 	HAL_GPIO_WritePin(GPIOB, gpio_pin, !!remote_param);
