--- gophercannon/automation/build_networks.sh
@@ -1,4 +1,0 @@
-TRANSIENT_DURATION=2
-WOBBLE_AMOUNT=1
-
-python3 gophercannon.py networks/go4-22c.yaml -r -n -s -t $TRANSIENT_DURATION -w $WOBBLE_AMOUNT
