--- gophercannon/reports/js/scripts.js
@@ -1,1 +1,0 @@
-// Empty JS for your own code to be here
