+++ gophercannon/requirements.txt
@@ -1,17 +1,24 @@
+astroid==2.4.2
 cycler==0.10.0
 gitdb==4.0.5
-GitPython==3.1.12
+GitPython==3.1.11
+isort==5.6.4
 Jinja2==2.11.2
-json2html==1.3.0
 kiwisolver==1.3.1
+lazy-object-proxy==1.4.3
 MarkupSafe==1.1.1
 matplotlib==3.3.3
+mccabe==0.6.1
 munch==2.5.0
-numpy==1.19.5
-Pillow==8.1.0
+numpy==1.19.4
+Pillow==8.0.1
+pylint==2.6.0
 pyparsing==2.4.7
 python-dateutil==2.8.1
 PyYAML==5.3.1
+rope==0.18.0
 six==1.15.0
 smmap==3.0.4
 tabulate==0.8.7
+toml==0.10.2
+wrapt==1.12.1
