--- gophercannon/templates/landing_page_TEMPLATE.html.jinja2
@@ -1,65 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-
-    <link rel='shortcut icon' type='image/x-icon' href='favicon.ico' />
-
-    <title>GopherCANnon Reports</title>
-
-    <meta name="description" content="Source code generated using layoutit.com">
-    <meta name="author" content="LayoutIt!">
-
-    <link href="css/bootstrap.css" rel="stylesheet">
-    <link href="css/style.css" rel="stylesheet">
-
-  </head>
-  <body>
-    <div class="container-fluid">
-        <div class="row">
-            <div class="col-md-12">
-                <div class="row">
-                    <div class="col-md-2">
-                    </div>
-                    <div class="col-md-8">
-                        <h1 class="text-center">
-                            Networks reports for
-                            <a href = "{{ commit_url }}">
-                                {{ repo_hash }}
-                            </a>
-                        </h1>
-                    </div>
-                    <div class="col-md-2">
-                    </div>
-                </div>
-                {% for url in urls %}
-                <div class="row">
-                    <div class="col-md-4">
-                    </div>
-                    <div class="col-md-4">
-                        <h4 class="text-center">
-                            <a href = "{{ url.url }}">
-                                {{ url.name }}
-                            </a>
-                        </h4>
-                    </div>
-                </div>
-                {% endfor %}
-                <div class="row">
-                    <div class="col-md-2">
-                    </div>
-                    <div class="col-md-8">
-                        <img alt="Bootstrap Image Preview" src="logo.jpg" class="rounded-circle"/>
-                        <br/>
-                    </div>
-                </div>
-            </div>
-        </div>
-    </div>
-    <script src="js/jquery.min.js"></script>
-    <script src="js/bootstrap.min.js"></script>
-    <script src="js/scripts.js"></script>
-  </body>
-</html>
