--- gophercannon/templates/report_TEMPLATE.html.jinja2
@@ -1,98 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-
-    <link rel='shortcut icon' type='image/x-icon' href='favicon.ico' />
-
-    <title>{{ pretty_name }} Report</title>
-
-    <meta name="description" content="Source code generated using layoutit.com">
-    <meta name="author" content="LayoutIt!">
-
-    <link href="css/bootstrap.css" rel="stylesheet">
-    <link href="css/style.css" rel="stylesheet">
-
-  </head>
-  <body>
-    <div class="container-fluid">
-        <div class="row">
-            <div class="col-md-12">
-                <div class="row">
-                    <div class="col-md-4">
-                    </div>
-                    <div class="col-md-4">
-                        <h1 class="text-center">
-                            {{ pretty_name }} Report
-                        </h1>
-                    </div>
-                    <div class="col-md-4">
-                    </div>
-                </div>
-                {% if steady_state != 0 -%}
-                <div class="row">
-                    <div class="col-md-2">
-                    </div>
-                    <div class="col-md-8">
-                        <h4 class="text-left">
-                            Static Analysis
-                        </h4>
-                        Bus utilization: {{ steady_usage }}%
-                    </div>
-                </div>
-                {% endif %}
-                {% if transient_duration > 0 -%}
-                <div class="row">
-                    <div class="col-md-2">
-                    </div>
-                    <div class="col-md-8">
-                        <h4 class="text-left">
-                            Transient Analysis
-                        </h4>
-                        Simulation duration: {{ transient_duration }}s
-                        <br/>
-                        Bus utilization over the final second of simulation: {{ transient_usage }}%
-                        <br/>
-                        {{ transient_table }}
-                        <img alt="Bootstrap Image Preview" src="{{ image_file }}" class="img-responsive"/>
-                    </div>
-                </div>
-                {% endif %}
-                <div class="row">
-                    <div class="col-md-2">
-                    </div>
-                    <div class="col-md-8">
-                        <h4 class="text-left">
-                            Network Documentation
-                        </h4>
-                        {{ network_table }}
-                    </div>
-                </div>
-                <div class="row">
-                    <div class="col-md-2">
-                    </div>
-                    <div class="col-md-8">
-                        <h4 class="text-left">
-                            Metadata
-                        </h4>
-                        {{ metadata_table }}
-                    </div>
-                </div>
-                <div class="row">
-                    <div class="col-md-2">
-                    </div>
-                    <div class="col-md-8">
-                        <img alt="Bootstrap Image Preview" src="logo.jpg" class="rounded-circle"/>
-                        <br/>
-                    </div>
-                </div>
-            </div>
-        </div>
-    </div>
-    <script src="js/jquery.min.js"></script>
-    <script src="js/bootstrap.min.js"></script>
-    <script src="js/scripts.js"></script>
-  </body>
-</html>
